+++ client/src/hooks/footer/info/useFooterInfoHook.ts
